backend/requirements.txt
openai==1.55.3
fastapi==0.115.6
python-dotenv==1.0.1
pydantic==2.10.4
uvicorn[standard]==0.34.0